dev-api/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Hariharan",email = "hariharankvasn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
package-mode = false
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "emails (>=0.6,<0.7)",
    "aiosmtplib (>=4.0.2,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "uuid7 (>=0.1.0,<0.2.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-mock (>=3.11.0,<4.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "psutil (>=5.9.0,<6.0.0)",
    "faker (>=15.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)"
]
